# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Trademark {
  id: number;
  date: string;
  prefix: string;
  clientNo: string;
  caseNo: string;
  /** @nullable */
  folderNo?: string | null;
  /** @nullable */
  appName?: string | null;
  /** @nullable */
  appClass?: string | null;
  /** @nullable */
  tmNo?: string | null;
  city: string;
  stage: string;
  /** @nullable */
  subStage?: string | null;
  status: string;
  isDuplicate: boolean;
  isTm11: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** Origin of the record: local or sheets */
  source: string;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface TrademarkInput {
  date: string;
  prefix: string;
  clientNo: string;
  caseNo: string;
  /** @nullable */
  folderNo?: string | null;
  appName: string;
  /** @nullable */
  appClass?: string | null;
  /** @nullable */
  tmNo?: string | null;
  city: string;
  stage: string;
  /** @nullable */
  subStage?: string | null;
  status: string;
  isDuplicate: boolean;
  isTm11: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
}

export interface TrademarkUpdate {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  prefix?: string | null;
  /** @nullable */
  clientNo?: string | null;
  /** @nullable */
  caseNo?: string | null;
  /** @nullable */
  folderNo?: string | null;
  /** @nullable */
  appName?: string | null;
  /** @nullable */
  appClass?: string | null;
  /** @nullable */
  tmNo?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  subStage?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  isDuplicate?: boolean | null;
  /** @nullable */
  isTm11?: boolean | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
}

export interface StageCount {
  stage: string;
  count: number;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface SubStageCount {
  subStage: string;
  count: number;
}

export interface TrademarkStats {
  total: number;
  byStage: StageCount[];
  byCity: CityCount[];
  duplicates: number;
  tm11Count: number;
  byNumericStage: StageCount[];
  byAssignedSubStage: SubStageCount[];
}

export interface SyncResult {
  synced: number;
  message: string;
}

export type ListTrademarksParams = {
/**
 * Search by TM No, app name, or folder number
 */
search?: string;
/**
 * Filter by stage
 */
stage?: string;
/**
 * Filter by city
 */
city?: string;
/**
 * Data source: local (DB only), sheets (Google Sheets only), all (merged)
 */
source?: ListTrademarksSource;
};

export type ListTrademarksSource = typeof ListTrademarksSource[keyof typeof ListTrademarksSource];


export const ListTrademarksSource = {
  local: 'local',
  sheets: 'sheets',
  all: 'all',
} as const;