# lib/api-zod/src/generated/types/subStageCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SubStageCount {
  subStage: string;
  count: number;
}