# lib/api-zod/src/generated/types/trademarkInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TrademarkInput {
  date: string;
  prefix: string;
  clientNo: string;
  caseNo: string;
  /** @nullable */
  folderNo?: string | null;
  appName: string;
  /** @nullable */
  appClass?: string | null;
  /** @nullable */
  tmNo?: string | null;
  city: string;
  stage: string;
  /** @nullable */
  subStage?: string | null;
  status: string;
  isDuplicate: boolean;
  isTm11: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
}