# lib/api-zod/src/generated/types/trademarkStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CityCount } from './cityCount';
import type { StageCount } from './stageCount';
import type { SubStageCount } from './subStageCount';

export interface TrademarkStats {
  total: number;
  byStage: StageCount[];
  byCity: CityCount[];
  duplicates: number;
  tm11Count: number;
  byNumericStage: StageCount[];
  byAssignedSubStage: SubStageCount[];
}