# lib/api-zod/src/generated/types/trademarkUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TrademarkUpdate {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  prefix?: string | null;
  /** @nullable */
  clientNo?: string | null;
  /** @nullable */
  caseNo?: string | null;
  /** @nullable */
  folderNo?: string | null;
  /** @nullable */
  appName?: string | null;
  /** @nullable */
  appClass?: string | null;
  /** @nullable */
  tmNo?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  subStage?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  isDuplicate?: boolean | null;
  /** @nullable */
  isTm11?: boolean | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
}